Extract fake repo_files
commit-id:b6265a26
